python-client/requirements.txt
python-socketio[client]==5.13.0
requests==2.32.3
